refactor(costs): split Costs.tsx into modular components, move data layer to react-query
Costs.tsx drops from 2599 to 763 lines. Types move to costs/types.ts, pure
helpers to costs/lib.ts, and the page body splits into six section
components (CostPulseHeader, MetricCards, AgentSplitCard, BudgetCard,
DailySpendSection, CostDriversSection). The hand-rolled fetch chain becomes
react-query: the /api/costs stale-retry loop is now a conditional
refetchInterval preserving the same fast_fallback/refreshing/stale
conditions, the four side fetches use the shared useApi hook, and the
budget save is a useMutation. Budget input now initializes once instead of
being clobbered by every background refresh. costsPageBehavior guards now
scan the new module files alongside Costs.tsx; assertions unchanged.
